mysql/mysql.go
package mysql

import (
	"errors"
	"strings"

	"github.com/mstgnz/sqlporter"
)

type MySQL struct {
	schema *sqlporter.Schema
}

// NewMySQL creates a new MySQL parser instance
func NewMySQL() *MySQL {
	return &MySQL{
		schema: &sqlporter.Schema{},
	}
}

// Parse parses MySQL dump content
func (m *MySQL) Parse(content string) (*sqlporter.Schema, error) {
	if content == "" {
		return nil, errors.New("empty content")
	}

	// TODO: MySQL dump parsing operations will be implemented here
	return m.schema, nil
}

// Generate generates MySQL dump from schema
func (m *MySQL) Generate(schema *sqlporter.Schema) (string, error) {
	if schema == nil {
		return "", errors.New("empty schema")
	}

	var result strings.Builder

	// TODO: MySQL dump generation operations will be implemented here

	return result.String(), nil
}
